backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
torch==2.1.1
python-dateutil==2.8.2
pydantic>=2.0.0
python-multipart==0.0.6
python-dotenv>=1.0.0
pulp>=2.7.0
entsoe-py>=0.6.1
